bugfix: affichage vampirisme equipement dans la fiche de perso

+++ fonctions_sql/f_perso_vampirisme.sql
@@ -12,7 +12,7 @@ declare
 
   personnage alias for $1;
   code_retour numeric;
-  v_vampirisme integer;
+  v_vampirisme numeric;
 
 begin
   code_retour := 0;
